25 create alternative simulation function (#32)
Modified the core simulation function to take a Boolean parameter that defaults to True. If set to false, the simulation will not store any data about the blue path. The blue path list will just contain the start and end position. This will make iterating the simulation less memory intensive.